+++ ethergo/submitter/submitter_test.go
@@ -2,6 +2,8 @@ package submitter_test
 
 import (
 	"fmt"
+	"math/big"
+
 	"github.com/brianvoe/gofakeit/v6"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/core/types"
@@ -20,7 +22,6 @@ import (
 	dbMocks "github.com/synapsecns/sanguine/ethergo/submitter/db/mocks"
 	submitterMocks "github.com/synapsecns/sanguine/ethergo/submitter/mocks"
 	"github.com/synapsecns/sanguine/ethergo/util"
-	"math/big"
 )
 
 func (s *SubmitterSuite) TestSetGasPrice() {
@@ -70,6 +71,39 @@ func (s *SubmitterSuite) TestSetGasPrice() {
 
 	// 4. Test with bump (TODO)
 	// 5. Test with bump and max (TODO)
+}
+
+func (s *SubmitterSuite) TestGetGasBlock() {
+	wall, err := wallet.FromRandom()
+	s.Require().NoError(err)
+
+	signer := localsigner.NewSigner(wall.PrivateKey())
+
+	chainID := s.testBackends[0].GetBigChainID()
+	client := new(clientMocks.EVM)
+
+	cfg := &config.Config{}
+	ts := submitter.NewTestTransactionSubmitter(s.metrics, signer, s, s.store, cfg)
+	currentHeader := &types.Header{Number: big.NewInt(1)}
+
+	// 1. Test with failed HeaderByNumber RPC call; Error is expected.
+	mockErrMsg := "mock error"
+	client.On(testsuite.GetFunctionName(client.HeaderByNumber), mock.Anything, mock.Anything).Times(5).Return(nil, fmt.Errorf(mockErrMsg))
+	gasBlock, err := ts.GetGasBlock(s.GetTestContext(), client, int(chainID.Int64()))
+	s.Nil(gasBlock)
+	s.NotNil(err)
+
+	// 2. Test with successful HeaderByNumber RPC call.
+	client.On(testsuite.GetFunctionName(client.HeaderByNumber), mock.Anything, mock.Anything).Once().Return(currentHeader, nil)
+	gasBlock, err = ts.GetGasBlock(s.GetTestContext(), client, int(chainID.Int64()))
+	s.Require().NoError(err)
+	s.Equal(gasBlock.Number.String(), currentHeader.Number.String())
+
+	// 3. Test with failed HeaderByNumber RPC call; the cached value should be used.
+	client.On(testsuite.GetFunctionName(client.HeaderByNumber), mock.Anything, mock.Anything).Times(5).Return(nil, fmt.Errorf(mockErrMsg))
+	gasBlock, err = ts.GetGasBlock(s.GetTestContext(), client, int(chainID.Int64()))
+	s.Require().NoError(err)
+	s.Equal(gasBlock.Number.String(), currentHeader.Number.String())
 }
 
 func (s *SubmitterSuite) TestGetNonce() {