+++ packages/contracts-core/test/suite/hubs/ExecutionHub.t.sol
@@ -7,6 +7,7 @@ import {
     GasLimitTooLow,
     GasSuppliedTooLow,
     IncorrectDestinationDomain,
+    IncorrectOriginDomain,
     IncorrectMagicValue,
     IncorrectSnapshotRoot,
     MessageOptimisticPeriod,
@@ -104,6 +105,31 @@ abstract contract ExecutionHubTest is AgentSecuredTest {
         bytes memory rcptPayload =
             verify_messageStatus(msgLeaf, snapRoot, sm.rsi.stateIndex, MessageStatus.Success, executor, executor);
         verify_receipt_valid(rcptPayload);
+    }
+
+    function test_execute_base_recipientEOA(Random memory random) public {
+        recipient = random.nextAddress();
+        // Create some simple data
+        (RawBaseMessage memory rbm, RawHeader memory rh, SnapshotMock memory sm) = createDataRevertTest(random);
+        // Create messages and get origin proof
+        RawMessage memory rm = createBaseMessages(rbm, rh, localDomain());
+        msgPayload = rm.formatMessage();
+        msgLeaf = rm.castToMessage().leaf();
+        bytes32[] memory originProof = getLatestProof(rh.nonce - 1);
+        // Create snapshot proof
+        adjustSnapshot(sm);
+        (bytes32 snapRoot, bytes32[] memory snapProof) = prepareExecution(sm);
+        // Make sure that optimistic period is over
+        uint32 timePassed = random.nextUint32();
+        timePassed = uint32(bound(timePassed, rh.optimisticPeriod, rh.optimisticPeriod + 1 days));
+        skip(timePassed);
+        vm.expectEmit();
+        emit Executed(rh.origin, msgLeaf, false);
+        vm.prank(executor);
+        testedEH().execute(msgPayload, originProof, snapProof, sm.rsi.stateIndex, rbm.request.gasLimit);
+        bytes memory rcptPayloadFirst =
+            verify_messageStatus(msgLeaf, snapRoot, sm.rsi.stateIndex, MessageStatus.Failed, executor, address(0));
+        verify_receipt_valid(rcptPayloadFirst);
     }
 
     function test_execute_base_recipientReverted_thenSuccess(Random memory random) public {
@@ -257,6 +283,27 @@ abstract contract ExecutionHubTest is AgentSecuredTest {
         verify_messageStatusNone(msgLeaf);
     }
 
+    function test_execute_base_revert_optimisticPeriodMinus1Second() public {
+        Random memory random = Random("random salt");
+        // Create some simple data
+        (RawBaseMessage memory rbm, RawHeader memory rh, SnapshotMock memory sm) = createDataRevertTest(random);
+        // Create messages and get origin proof
+        RawMessage memory rm = createBaseMessages(rbm, rh, localDomain());
+        msgPayload = rm.formatMessage();
+        msgLeaf = rm.castToMessage().leaf();
+        vm.assume(rh.optimisticPeriod != 0);
+        bytes32[] memory originProof = getLatestProof(rh.nonce - 1);
+        // Create snapshot proof
+        adjustSnapshot(sm);
+        (, bytes32[] memory snapProof) = prepareExecution(sm);
+        // Make sure that optimistic period is NOT over
+        uint32 timePassed = rh.optimisticPeriod - 1;
+        skip(timePassed);
+        vm.expectRevert(MessageOptimisticPeriod.selector);
+        vm.prank(executor);
+        testedEH().execute(msgPayload, originProof, snapProof, sm.rsi.stateIndex, rbm.request.gasLimit);
+    }
+
     function test_execute_base_revert_gasLimitTooLow(Random memory random) public {
         // Create some simple data
         (RawBaseMessage memory rbm, RawHeader memory rh, SnapshotMock memory sm) = createDataRevertTest(random);
@@ -324,6 +371,21 @@ abstract contract ExecutionHubTest is AgentSecuredTest {
         vm.prank(executor);
         testedEH().execute(msgPayload, originProof, snapProof, sm.rsi.stateIndex, rbm.request.gasLimit);
         verify_messageStatusNone(msgLeaf);
+    }
+
+    function test_execute_base_revert_originSameDomain() public {
+        RawBaseMessage memory rbm;
+        rbm.sender = rbm.recipient = addressToBytes32(recipient);
+        RawMessage memory rm;
+        rm.header.flag = uint8(MessageFlag.Base);
+        rm.header.origin = localDomain();
+        rm.header.destination = localDomain();
+        rm.header.nonce = 1;
+        rm.header.optimisticPeriod = 1 seconds;
+        rm.body = rbm.formatBaseMessage();
+        msgPayload = rm.formatMessage();
+        vm.expectRevert(IncorrectOriginDomain.selector);
+        testedEH().execute(msgPayload, new bytes32[](0), new bytes32[](0), 0, 0);
     }
 
     function test_execute_base_revert_reentrancy(Random memory random) public {